# backend/src/types/order.ts
import { components } from "../schemas";

export type OrderDto = components['schemas']['Order'];
export type OrderRequestBodyDto = components['schemas']['OrderRequestBody'];

export type OrderItemDto = components['schemas']['OrderItem'];
export type OrderItemRequestBodyDto = components['schemas']['OrderItemRequestBody'];