# backend/src/types/user.ts
import { components } from "../schemas";

export type UserDto = components['schemas']['User'];
export type RegisterRequestBodyDto = components['schemas']['RegisterRequestBody'];
export type LoginRequestBodyDto = components['schemas']['LoginRequestBody'];